feat: Enable single minting in the Studio (#842)
* feat: Enable single minting in the Studio

* fix: type

* fix: review
